Add doc listing Conscrypt capabilities.
This will be useful to indicate what we do and don't support, as well
as a place to indicate our implementation-specific choices (like MGF1
digest or TLS 1.3 choices).

Also standardize on lowercase "with" in Signature algorithm IDs.  The
RSA ones used "With" and the ECDSA and PSS ones used "with", so it's
2:1 in favor of lowercase.  JCA algorithm identifiers are matched
case-insensitively, so this is just for display purposes.