Extract loading of default CA store to Platform.
The default CA store on Android is built into the platform, so we can
simply fetch it.

For OpenJDK, we look for the highest-priority non-Conscrypt
TrustManagerFactory that produces at least one certificate.  If the
user has just installed Conscrypt as the first provider, which is
typical, this means we should get the OpenJDK-provided implementation,
which uses a variety of system properties to determine where to find
the CA certs.  Otherwise, whatever system they've previously been
using to choose the set of trusted CA certs will continue to work.