Fix Android benchmark references (#508)
The Android benchmarks were assuming a lot of things will stay in the
same place. Instead of hardcoding paths, try to use project properties
to discover where the files are.

Additionally make sure all the tasks have dependencies in the right
Gradle lifecycle so they can be run individually without causing an
error.