Upgrade wrapper to Gradle 4.8.1 (#506)
* Upgrade wrapper to Gradle 4.8.1

This fixes the temporary hack for Windows 10 SDK detection. Also updates
some build files to work with changes in 4.8.1.

This necessitates the upgrade of Android SDK build tools which causes
cascading changes being needed in other dependencies and build files in
Android.

* Apply the Android Tools in the root project

Otherwise each subproject will have its own copy of the class in its own ClassLoader. This breaks
Android Lint tool which assumes it can cast directly to Class type.

* Remove custom Test task settings

These do not appear to be needed since they are not set in the regular
test stanza. They should be set automatically by virtue of this being
a Test type task in a java-library project.