Provide TrustManagerFactory.
This is necessary for users who want to enable TLS 1.3, since the
TrustManagerFactory implementation shipped with OpenJDK throws an
exception if it encounters an SSLSocket or SSLEngine that's negotiated
TLS 1.3.  Use our TrustManager in tests.

Also adds a HostnameVerifier in the tests that does the simplest
thing, because our TrustManager verifies hostnames by default, whereas
the OpenJDK one doesn't, and the bundled HostnameVerifier on OpenJDK
just fails to verify anything.