Merge ConscryptSocketBase and AbstractConscryptSocket
We used to need this separation because we had delegating sockets that
inherited from AbstractConscryptSocket but not ConscryptSocketBase,
but those are gone now, so we no longer need to have two classes here.

While the remaining file is named AbstractConscryptSocket (to parallel
AbstractConscryptEngine), what really happened is the abstract methods
from AbstractConscryptSocket got merged into ConscryptSocketBase, and
then the resulting class was renamed.