Add TLS 1.3 benchmarks
Adds a BenchmarkProtocol parameter to ClientSocketBenchmark and
EngineHandshakeBenchmark to determine which protocol to use.

Switched EngineHandshakeBenchmark off TestUtils.doEngineHandshake and
onto its own implementation.  This lets us simulate network transit
time as well as eliminating all the test assertions and generally
operating closer to the typical usage of an SSLEngine.

Also adds the ability to specify JMH params on the command line.