Fix X509_verify exceptions (#537)
Unexpectedly, X509_verify does not set an entry in the error queue
when the signature algorithm in a certificate's info doesn't match the
algorithm in the signature block, it just returns 0.  This trips an
AssertionError in throwExceptionFromBoringSSLError, so check for that
condition explicitly.

Also default to throwing CertificateException instead of
RuntimeException when X509_verify encounters an error that we haven't
explicitly handled.  I don't know if there are any such errors in
practice, but it'll be more correct if there are.