Allow d2i_X509_bio and friends to not set an error (#552)
BoringSSL recently changed so that d2i_X509_bio doesn't set an error
in the error queue when it receives certain garbage inputs.  We don't
actually care what error is set in the queue (we just end up catching
whatever is thrown and throwing CertificateException), so change to
tolerate having no error in the queue.