Revert "Prefer Libcore.os over Os outside of android.* packages. (#358)"
The original CL was submitted because it was thought (at the time) that
android.system would become inaccessible to Conscrypt; the opposite is
now the case: We intend to make Libcore.os private within libcore, so
it will become inaccessible to Conscrypt. Therefore this CL reverts back
to the previous use of android.system.Os.

This reverts commit c4b87c4390c480a0b17be9ad99b0f5a1fb7559c1.