Update handling of empty error queue (#555)
Some BoringSSL functions don't make any guarantees about the state of
the error queue if they return an error, and in most cases we don't
actually care what error occurred.  Instead of throwing AssertionError
when we encounter an error with no entry in the error queue, throw the
default exception type (which is usually RuntimeException).

Also removes NativeCrypto.ERR_peek_last_error(), which was only used
in tests.  We now have the CHECK_ERROR_QUEUE_ON_RETURN macro that
serves the same purpose but does it more comprehensively.

Fixes #553.