Switch appveyor.yml from Yasm to NASM.
NASM is actively maintained while Yasm appears to be abandoned these
days. Both should work for now, but we've switched to recommending
building BoringSSL with NASM. Chromium is also using NASM to build
BoringSSL now. (Yasm may break in the future if we add assembly that
uses some Intel instructions that aren't in Yasm.)

Happily, chocolatey's NASM package is up-to-date (last update a week
ago), unlike its Yasm package, which is one version behind. So this
means appveyor.yml is simpler now.