Make OpenSSLBIOInputStream read as fully as possible
Some of the callers of BIO APIs in BoringSSL (particularly the DER
decoder) depend on the BIO read returning either a full buffer or an
EOF, but InputStream allows returning less than the requested amount
of data without hitting EOF.  This causes the DER decoder to fail in
some cases (specifically, we've seen it fail with overly large OCSP
responses where the socket hasn't buffered the entire response before
we try to decode it).

Fix this by repeatedly calling read() until our output buffer is full
or EOF is reached.

Also add a test for decoding DER-encoded certs, we previously only
tested PEM-encoded ones.