Create ActiveSession during transitionTo() (#586)
The SessionContext we should use in ActiveSession depends on whether
the engine is in client or server mode, which we don't know until the
handshake is started, so wait until handshake start to create it.  As
a bonus, this makes the comment for activeSession correct again.

Also make closeInbound() the same as closeOutbound() where if you
close it before the handshake starts it just closes the engine
entirely.  STATE_CLOSED_INBOUND is assumed to be a post-handshake
state by a bunch of stuff, so we need to skip it.

Fixes #585.