Clean up session handling
The session passed to the SSLSessionBindingListener was always an
ExternalSession, even when that session was wrapped to support
ExtendedSSLSession, which meant that observers saw two different
sessions.  Refactor putValue() and removeValue() on ExternalSession to
take the session so that the wrappers can provide themselves.  Add a
test for SSLSessionBindingListener calling.

Slightly clean up ConscryptEngineSocket.getSession().  It doesn't need
to do cache the session from before it tries to handshake, and doing
so means that if you call getSession() twice in succession it might
return different results for no good reason.

Get rid of SSLNullSession.isNullSession, SSLUtils.unwrapSession, and
SessionDecorator, since that entire chain was only used in
ConscryptEngineSocket.getSession().

Fixes #593