Make ConscryptEngineSocket call correct methods (#643)
ConscryptEngineSocket uses an SSLEngine internally to function, and
that naturally calls the SSLEngine-accepting methods on
X509ExtendedTrustManager to do trust checks.  We were passing our
given trust manager directly to the SSLEngine, which resulted in an
SSLSocket implementation that ended up calling SSLEngine-based
methods, which isn't right.

Instead, create a delegating trust manager that maps the calls to the
correct objects.  On platforms where X509ExtendedTrustManager isn't
available, we can pass in the provided trust manager directly, since
it doesn't receive a reference to the calling object.

Also adds tests for getHandshakeSession() that ensure it functions in
the middle of the handshake and provides properties that should be
set.