Add AES-GCM-SIV
AES-GCM-SIV is a nonce misuse-resistant variant of AES-GCM.  It's
probably more useful to recommend as a general-use AEAD than GCM
alone, since nonce management is something that's easy to screw up in
some contexts, and BoringSSL supports it, so expose it.

See https://tools.ietf.org/html/draft-irtf-cfrg-gcmsiv-09 for more
details.