Add ServiceTester helper class
We have a lot of tests that follow the same basic pattern of "for all
providers, for some/all algorithms of a given type, check {X, Y, Z}".
This introduces a helper object that handles the common parts (finding
supported providers, matching errors to providers, etc) so the tests
can focus on the test logic.

This change also cleans up a bunch of minor test problems:

* Test arrays with assertArraysEqual(), not assertTrue(Arrays.equal())
* Unusual formatting (extraneous blocks, etc.)
* Unnecessary installing of BouncyCastleProvider in TestKeyStore