Fetch local certs when available
Local certificates don't change once they're set, so we don't have to
wait for any particular point to fetch and cache them.  If someone
requests them, we can just get them from the SSL if available.

Also adds tests for getHandshakeSession() on the server side, which is
a time when these certs would previously not have been available.

Fixes #634