Fix ConscryptEngine state machine (#687)
The SSLEngine state machine is slightly more complicated than our
previous code presumed.  After closeInbound() or closeOutbound() are
called, isInboundDone() and isOutboundDone() shouldn't immediately
return true.  Rather, closeInbound() and closeOutbound() indicate the
caller's intention to never supply any more input data in the given
direction, but isInboundDone() and isOutboundDone() indicate that no
more output data in the given direction will be produced.  This means
those methods need to be adjusted to take into account the state of
internal plaintext/ciphertext buffers.

This simplifies and corrects some of the other code as well.  In
particular, we no longer have to stash exceptions that are thrown
during the handshake, because the caller is expected to see those
exceptions and drain the outgoing buffer (which previously they had no
way of doing).

Fixes #577