Fix engine socket tests
The engine test system property was being overwritten, which resulted
in the engine socket test not actually being executed with the engine
socket.  This fixes that problem, so :testEngineSocket now actually
runs with the engine socket.

This revealed several test failures that were all due to the engine
and engine socket closing behavior.  In particular, the engine
wouldn't mark itself as closed when an SSLException was thrown, which
it is supposed to per the documentation.  This meant that when the
engine socket attempted to flush the outgoing buffer, it would just
trigger whatever exception had originally caused the problem again,
the outgoing buffer wouldn't get flushed, and alerts wouldn't reach
the peer.  Closing the engine whenever an exception propagates out
fixes this problem, along with tolerating a CLOSED status return in
the engine socket.