Move openjdk-integ-tests to common (#720)
The tests in openjdk-integ-tests haven't matched their name in a
while: they test Conscrypt-specific features in places, we run them on
platforms other than OpenJDK, and they're a combination of
integration, unit, and regression tests.  To ease confusion, move them
into common alongside other common tests.  Their package names (eg,
org.conscrypt.java.security) will still indicate that they're testing
the overall interface rather than Conscrypt-specific classes, but it
should be more clear that they're just part of the overall testing
scheme of Conscrypt on all platforms.

As part of this, remove the architecture-specific test rules in
openjdk and only test on the preferred architecture.  The additional
test classes increase the running time of the test rule to multiple
minutes, and neither the new nor the old tests are obviously
architecture-sensitive. In the rare case where it's desired to test on
multiple architectures, we can run the test rule twice in different
configurations.

We also change the existing OpenJDK tests to run in a suite that
installs Conscrypt before running.  This works better with the new
tests and makes the tests not need to manage provider installation
themselves.  As part of this, also remove calls to TestUtils methods
that need the Conscrypt provider to be installed during static
initialization; during static initialization, the provider isn't
installed and having those calls makes TLS tests fail on OpenJDK < 11.