Do not include tools.jar hack for doclet
Finding the correct tools is handled by Gradle now, so this is not
needed. At any rate, JDK 9 removed the tools.jar and split it up and
there is no single file to find the tools. It must be found through
an API that respects modules.

This may not be necessary to even include in Gradle 5.x.