Migrate to using maven-publish plugin
The old Gradle "maven" plugin has been deprecated and the new way of publishing
is "maven-publish." Switch to that in order to be able to update to newer versions
of Gradle. The plugins supporting Android with Gradle publishing have stopped
development, so upgrading to a newer Gradle version becomes difficult without
doing this first.