Avoid holding a GlobalRef to ALPN selector
The ALPN selector code used to hold a GlobalRef for the callback. This
appeared to make it impossible to discard the ref since there wasn't any
trigger for the AppData to be cleared as there is with the
SSLHandshakeCallbacks. Instead move the management of the reference into
Java and use the same SSLHandshakeCallbacks to manage calling into the
ALPN selector code.

This fixes #711