Use GitHub Actions for building
GitHub Actions allows us to build on all three supported platforms:
Linux, MacOS, and Windows. We can also use the artifacts together to
test building the Uber JAR flow. Previously we had no way to do this in
a CI environment.

This should allow the instructions to be tested on every change instead
of only when we are preparing a release.