Skip some tests on Zulu release of OpenJDK
It appears that some parts of Java 9 JCE were backported by Azul Systems for their
Zulu release of OpenJDK. However, the actual implementation was not included so
trying to test those algorithms results in a NoClassDefFound error. Skip them until
the bug is addressed in a subsequent release of Zulu.