Fix the MixedMutabilityReturnType errors
For internal APIs we can return mixed mutability, but it would preferable to mark
this as immutable always. It appears that ErrorProne assumes you have Guava available
to use the ImmutableList class.

For external APIs we should probably always return mutable types to avoid errors
where the caller expects to be able to do .add(...) on it.