Ignore unused methods in test class
This class has several fields that are not currently used, but could be in the future.
They are kind of hard to find without the methods, so just mark this whole class as
unused suppression so we do not have to dig around to resurrect them later if they are
used.