Make SSLSocketVersionCompatibilityTest not flaky
This test checks that SSLHandshakeException messages are due to connection closed,
but the internal SSLUtils utility was being used to convert the EOFException to a
SSLHandshakeException by reusing the message from EOFException. Since we were not
using the EOFException constructor with a message, it was getting NPE when the test
got that exception.

Add the message that explains why an EOFException was thrown.