Update Mockito and Google Truth versions
Older version of Mockito use a method of creating mock classes that Java 11 does not
work correctly with. New versions of Mockito have a method that is compatible with old
and new versions of Java language.

Update calls to deprecated Matchers methods to use ArgumentMatchers for the new version
of Mockito as well.