Don't loop forever in writeInternal if making no progress.
PR #707 made ConscryptEngineSocket.SSLOutputSteam.writeInternal tolerant
of the underlying engine being CLOSED in order to ensure any pending
alerts get flushed to the underlying socket.

However in the case where SSLSocket.close() was called from another
thread causing the underlying socket to get closed too then this could
loop forever.  Added a condition that if no new data is produced for
sending in this state then signal the SSLSocket is closed by throwing
a SocketException, which is consistent with the
ConscryptFileDescriptorSocket behaviour.

This could also potentially be fixed ConscryptEngine.wrap() by throwing
directly from there instead of returning a CLOSED result, but I am not
convinced that is the only code path that can provoke this, so better
to ensure the writeInternal() loop will always exit.

The added test for this isn't a perfect regression test, as it's timing
dependent whether the writing thread will be looping in writeInternal
or blocked in socketOutputStream.write(), hence the test is only run
with autoClose=true until b/147870945 is resolved (Engine-based socket
doesn't wake blocked threads on close).  I'll rationalize the tests as
part of the fix for that, but this change is useful on its own as it
should unblock wider testing of the engine base socket again.