Allow selection of fd-based SSLSocket using a flag file.
This is a bit of an ugly hack, but temporary in nature and
less ugly that having to write a custom CTS test runner to
pass in a Java property.

Allows us to create a CTS test suite that exercises a
number of Conscrypt and non-Conscrypt APIs with the SSLSocket
implementation set to ConscryptFileDescriptorSocket even though
the AOSP default is ConscryptEngineSocket.

Flag file name chosen to minimise potential for accidental
collisions.

To be kept until ConscryptFileDescriptorSocket is removed in
order to ensure we continue to test both implementations on
all platforms.