Propagate socket implementation type changes to default factories.
Conscrypt.setUseEngineSocketByDefault() sets the default
SSLSocket implementation to use by calling the same method on
OpenSSLSocketFactoryImpl and OpenSSLServerSocketFactoryImpl.

Those methods set a static boolean which is used to determine
the initial default for new factories but which has no effect
on existing factories.  This is mostly what's wanted except the
default SSLSocketFactory (and SSLServerSocketFactory) can be
long lived and created by static initializers before any code
has a chance to call Conscrypt.setUseEngineSocketByDefault().

This change also change the flag for the default
SSLSocketFactory and SSLServerSocketFactory, so that an app
using these behaves as expected.

Negative:  This will change means that calling
Conscrypt.setUseEngineSocketByDefault will create a default
SSLServerSocketFactory, which some apps may never use, however
the overhead is low and it seems better than changing the
API to add a setUseEngineServerSocketByDefault() method,
as existing apps might need changing to call both methods.
Also this API should only last in the short to medium term
until we can retire the fd-based implementation.