Fix ConscryptEngineSocket.close().
Previoly, this method closed the underlying socket first and then
the SSLEngine.  However closing a connected SSLEngine queues a TLS
close notification which obviously can't be sent if the socket is
closed.  Also, the pending bytes prevent the engine from freeing
its native resources including pipe file descriptors until the
SSLEngine is eventually garbage collected.

Fixing this exposed that the fix for #781 was incomplete and relied
on the native SSL data *not* being cleared on close and so that
is also fixed herein.

This may also help with #835 but didn't help me to reproduce that.