Remove redundant X509_ALGOR_cmp check.
This check was added in be510b32477d98f12d1cbd2a9bf51d7c7e035719, to
work around some error-handling issues because, while BoringSSL did this
check, did not add an error to the error queue and Conscrypt's error
handling was very sensitive to this.

Since then, BoringSSL has fix the error-reporting, and Conscrypt has
fixed its error-handling to be based on the return value. This logic is
no longer necessary. (The test still passes when it is removed.)