Extract the SAN list with X509_get_ext_d2i.
This mirrors the issuer alternate names logic. While OpenSSL does cache
the SAN list on x->altname, they did not add an accessor for it.
Instead, it seems the public API in OpenSSL is now to extract it anew.

This removes the need for calling X509_check_ca to populate the weird
extensions cache.