Don't check X509 internals in get_*_ext_oids.
This is effectively a revert of #380. It should be impossible for
x509->cert_info and crl->crl to be NULL, short of a memory error or code
messing with structs. Also the other X509 functions don't check for
NULL, so any such object would crash elsewhere anyway.

We're looking to make the various X509 structs opaque, to align with
OpenSSL, at which point these fields will be inaccessible.