Update jarsigner logic for new Maven publishing plugin.
The previous hook for signing jars before Maven PGP signatures
were generated was not actually being called.

Additionally, it was signing only the generic jar, not the
platform one which actually gets uploaded.  This may have worked
with previous combinations of Gradle/Maven publishing, but
currently the platform jar is not added to
`configurations.archives.allArtifacts` (possibly a separate bug).
Workaround: search all artifacts to be uploaded for candidate
jars.

Tested on Macos and Linux.