Make testing of handshake session cipher suite consistent.
These tests should be agnostic about which suite is selected
provided that it is in the enabled set, however when the test was
duplicated into client and server versions it became opinionated
again. Similarly the error message became inconsistent during
upstreaming from AOSP.

Changed use of `final boolean[1]` to the preferred
`AtomicInteger` while I was there.

Test: gradlew check