Fix pathlen handling given other invalid extensions.
Due to weird behaviors in OpenSSL's API,
c221a4d1afd6a9601138fbcb000dba34a255a05a inadvertently changed
getBasicConstraints() behavior when an unrelated extension had a syntax
error. This tripped a broken test in Android.

The right thing to do here is complex and relates to when to raise
which syntax errors in an invalid certificate. For now, restore the old
behavior by extracting the basicConstraints extension separately. This
brings up the broader question of how to handle invalid certificates, so
I've left a host of TODOs for #916.