Refactor JNI BIO ref null checking.
Throwing on null is no longer optional in `to_SSL_BIO()`, i.e.  it
is always an error to pass a null BIO reference to JNI code.  All
existing usages worked in this mode anyway.

Rename `to_SSL_BIO()` to `to_BIO()` and use it for for all other
cases of passing BIO refs into JNI.  This is a slight behaviour
change for some methods as they will now throw NPE rather than some
BoringSSL parse exception but seems more correct.

If this CL is approved I'll follow on with similar changes for other
reference type so that the rule everywhere becomes: If a null
reference is acceptable then that case should be handled in Java.
If the reference is passed into JNI it should always trigger an NPE.