Fix ConscryptEngine.wrap() behaviour when multiple buffers passed in.
Fixes #929.

Behaviour unchanged for the common case of passing a single
buffer to `wrap()`. TODO(prb) Add a test for this.

When passing an array of small buffers, instead of passing them
individually to BoringSSL, which generates one TLS record per buffer,
coalesce theem into a single larger buffer so that a single maximal
sized TLS record is generated, which was the original intent here.

This is a slight behaviour change for apps which aren't hitting #929
but I believe based on commentary on the bug this is closer to
the RI behacviour.

Currently, if an app passes multiple buffers into a single `wrap()`
call, one TLS record is generated for each which means on the
receiving end multiple `unwrap()` calls are required so the app has
to loop, unwrapping until no more records are available.  After
this change only a single unwrap is required so the same code will
still work. However if an app is making assumptions about where TLS
record boundaries lie then it might have issue - such an app is
buggy as `SSLEngine` makes no guarentees in this area.  This assumption
was inadvertently baked into a couple of tests - removed it.