KeyAgreement.XDH for X25519 (#950)
* KeyAgreement.XDH for X25519

This implements the KeyAgreement.XDH for use with X25519 (aka
Curve25519) keys. It also implements the requisite KeyFactory.XDH and
KeyPairGenerator.XDH.

It does not currently inherit from the XECPrivateKey and XECPublicKey
classes. This will work in OpenJDK 11 and later, but will require some
creative compilation which will be done in a later commi

* Skip KeyPairGenerator.XDH for OpenJDK 11

There is a bug in OpenJDK 11 that encodes the private key in the
wrong way. It is fixed in JDK 12, but JDK 11 was kept broken to maintain
compatibility.