Rework TBSCertificate (re)encoding bindings. (#992)
An X.509 certificate contains of a TBSCertificate ("to be signed"), a
signature algorithm, and a signature. Although, in theory, DER
guarantees that re-encoding a parsed TBSCertificate gives back the same
bytes, it is safer and more robust to verify the encoded TBSCertificate
directly. (Also OpenSSL's parser is very lax and definitely not DER.)

Conscrypt needs the encoded TBSCertificate in two places: first to
implement getTBSCertificate(), used in signature verification, and
second as part of removing the CT poison extension in precertificate
handling.

OpenSSL's X509 type handles TBSCertificate a little strangely. It has a
cached encoded TBSCertificate that is populated on parse. But, to
support certificate issuance, one can also create an empty X509 object
and mutate it to make a certificate. These two use cases don't combine
very well; mutating a certificate silently puts the cached encoding in
an inconsistent state.

Upstream OpenSSL, when making X509 opaque, decided not to expose the
TBSCertificate or the cached encoding directly (X509_CINF or cert_info).
I think that was probably the right move as the cached data is a little
weird. Also X509_CINF_set_modified changes the thread-safety properties
of the X509 which seems especially poor. For CT, they have
i2d_X509_re_tbs which bundles together extracting the TBSCertificate and
forcing a re-encode. I've generalized it to also have i2d_X509_tbs,
which extracts TBSCertificate without a re-encode.

This PR switches Conscrypt to these APIs. Specifically:

- getTBSCertificate() uses the non-re-encoding version, because, except
  where it is called in CertificateEntry, it wants the original
  encoding.

- The CertificateEntry logic calls a new
  getTBSCertificateWithoutExtension() method. This method bundles
  together all the steps.

- getTBSCertificateWithoutExtension() has a single corresponding
  NativeCrypto method. This means that, while we still internally have
  an X509 object in the funny inconsistent state, it is never exposed to
  Java or OpenSSLX509Certificate. (If the modified bit is set on an
  X509, serializing it is no longer thread-safe.)

- Now unused bindings are removed. I've also renamed
  get_X509_cert_info_enc to get_X509_tbs_cert since "cert_info" no
  longer appears in the public API.