Bump Java language level to 1.8.
Only affects non-Android builds.  Android builds are still desugared
to be compatible with 1.7 by the SDK in order to run on older desserts.

We announced in the last release that we would be dropping support for 1.7.

Note that the Android-Stub build config previously had a comment that
it needed to be binary compatible with Android minSdkVersion. That
was erroneous, it's used for non-Android builds and so should shouldn't
redefine the source or target language level.

Tested:
./gradlew check
./gradlew conscrypt-android:check