Revise GENERAL_NAME_to_jobject. (#1032)
This fixes a few issues. First, the comment is a little confusing. The
value is an IA5String. The issue isn't whether it might accidentally be
a T61String, but that its contents may be invalid for an IA5String.

Next, I've replaced ASN1_PRINTABLE_type with a loop. ASN1_PRINTABLE_type
does actually work, but it's normally used in OpenSSL to auto-detect a
string type for a Latin-1 string, which is a little confusing.

Finally, I've tweaked this so it does not rely on the ASN1_STRING being
NUL-terminated. The ASN1_STRING is almost always NUL-terminated, but
best not to rely on this. (See CVE-2021-3712.)